requirements.txt
pydantic
pydantic-settings
python-dotenv
google-adk